게시판 CRUD

+++ src/main/java/com/prj1/controller/BoardController.java
@@ -3,10 +3,12 @@ package com.prj1.controller;
 import com.prj1.domain.Board;
 import com.prj1.service.BoardService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @Controller
@@ -21,8 +23,8 @@ public class BoardController {
     }
 
     @PostMapping("/add")
-    public String addPost(Board board, RedirectAttributes rttr) {
-        service.add(board);
+    public String addPost(Board board, Authentication authentication, RedirectAttributes rttr) {
+        service.add(board, authentication);
 
         rttr.addAttribute("id", board.getId());
         return "redirect:/board";
@@ -40,18 +42,22 @@ public class BoardController {
         return "board/view";
     }
 
+    // ?page=3
     @GetMapping("/")
-    public String home(Model model) {
+    public String home(@RequestParam(value = "page", defaultValue = "1") Integer page,
+                       Model model) {
         // 게시물 목록 조회(select)
         // 모델에 넣고
-        model.addAttribute("boardList", service.list());
+        model.addAllAttributes(service.list(page));
         // jsp로 포워드
         return "board/home";
     }
 
     @PostMapping("/delete")
-    public String delete(Integer id) {
-        service.remove(id);
+    public String delete(Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+        }
 
         return "redirect:/";
     }
@@ -66,8 +72,11 @@ public class BoardController {
     }
 
     @PostMapping("/modify")
-    public String modifyPost(Board board, RedirectAttributes rttr) {
-        service.modify(board);
+    public String modifyPost(Board board, Authentication authentication, RedirectAttributes rttr) {
+
+        if (service.hasAccess(board.getId(), authentication)) {
+            service.modify(board);
+        }
 
         rttr.addAttribute("id", board.getId());
         return "redirect:/board";
