+++ src/main/java/com/prj1/controller/MemberController.java
@@ -3,6 +3,7 @@ package com.prj1.controller;
 import com.prj1.domain.Member;
 import com.prj1.service.MemberService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,15 +14,14 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @Controller
 @RequiredArgsConstructor
-
-// /member/~~
 @RequestMapping("member")
-public class MemeberController {
+public class MemberController {
 
     private final MemberService service;
 
     @GetMapping("signup")
     public String signupForm() {
+
         return "member/signup";
     }
 
@@ -45,10 +45,12 @@ public class MemeberController {
     }
 
     @PostMapping("remove")
-    public String remove(Integer id) {
-        service.remove(id);
+    public String remove(Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+        }
 
-        return "redirect:/member/signup";
+        return "redirect:/logout";
     }
 
     @GetMapping("modify")
@@ -59,8 +61,10 @@ public class MemeberController {
     }
 
     @PostMapping("modify")
-    public String modify(Member member, RedirectAttributes rttr) {
-        service.modify(member);
+    public String modify(Member member, Authentication authentication, RedirectAttributes rttr) {
+        if (service.hasAccess(member.getId(), authentication)) {
+            service.modify(member);
+        }
 
         rttr.addAttribute("id", member.getId());
         return "redirect:/member";
@@ -72,5 +76,10 @@ public class MemeberController {
         System.out.println("email = " + email);
         String message = service.emailCheck(email);
         return message;
+    }
+
+    @GetMapping("login")
+    public String loginForm() {
+        return "member/login";
     }
 }
